pyproject.toml
[project]
name = "unitednationsofhaire"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "si>=0.1.1.post1",
    "six>=1.17.0",
]
